Fix a compilation issue with recent version of nanoflann

+++ src/structures/molecule.cpp
@@ -165,7 +165,7 @@ int Molecule::total_charge() const {
 std::vector<const Atom *> Molecule::get_close_atoms(const Atom &atom, double cutoff) const {
     std::vector<const Atom *> close_atoms;
 
-    std::vector<std::pair<size_t, double>> results;
+    std::vector<std::pair<uint32_t , double>> results;
     nanoflann::SearchParams params;
 
     auto matches_count = index_->radiusSearch(atom.pos().data(), cutoff * cutoff, results, params);
